requirements.txt
Django==1.6.1
Pillow==2.0.0
django-debug-toolbar==0.9.4
django-localeurl==1.5
django-modeltranslation==0.7.3
wsgiref==0.1.2
